# openapi.json
{"openapi":"3.1.0","info":{"title":"Model Registry API","description":"ML Model Registry for tracking, versioning, and discovering trained models","version":"1.0.0"},"paths":{"/api/v1/teams":{"get":{"tags":["teams"],"summary":"List Teams","operationId":"list_teams_api_v1_teams_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TeamList"},"type":"array","title":"Response List Teams Api V1 Teams Get"}}}}}},"post":{"tags":["teams"],"summary":"Create Team","operationId":"create_team_api_v1_teams_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teams/{team_id}":{"get":{"tags":["teams"],"summary":"Get Team","operationId":"get_team_api_v1_teams__team_id__get","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["teams"],"summary":"Delete Team","operationId":"delete_team_api_v1_teams__team_id__delete","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teams/{team_id}/models":{"post":{"tags":["models"],"summary":"Create Model","operationId":"create_model_api_v1_teams__team_id__models_post","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["models"],"summary":"List Team Models","operationId":"list_team_models_api_v1_teams__team_id__models_get","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ModelList"},"title":"Response List Team Models Api V1 Teams  Team Id  Models Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models":{"get":{"tags":["models"],"summary":"List All Models","operationId":"list_all_models_api_v1_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ModelList"},"type":"array","title":"Response List All Models Api V1 Models Get"}}}}}}},"/api/v1/models/{model_id}":{"get":{"tags":["models"],"summary":"Get Model","operationId":"get_model_api_v1_models__model_id__get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["models"],"summary":"Delete Model","operationId":"delete_model_api_v1_models__model_id__delete","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models/{model_id}/versions":{"post":{"tags":["versions"],"summary":"Create Version","operationId":"create_version_api_v1_models__model_id__versions_post","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelVersionCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelVersionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["versions"],"summary":"List Versions","operationId":"list_versions_api_v1_models__model_id__versions_get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ModelVersionList"},"title":"Response List Versions Api V1 Models  Model Id  Versions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models/{model_id}/versions/{version_id}":{"get":{"tags":["versions"],"summary":"Get Version","operationId":"get_version_api_v1_models__model_id__versions__version_id__get","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}},{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelVersionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["versions"],"summary":"Delete Version","operationId":"delete_version_api_v1_models__model_id__versions__version_id__delete","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}},{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/models/{model_id}/versions/{version_id}/stage":{"patch":{"tags":["versions"],"summary":"Update Version Stage","operationId":"update_version_stage_api_v1_models__model_id__versions__version_id__stage_patch","parameters":[{"name":"model_id","in":"path","required":true,"schema":{"type":"integer","title":"Model Id"}},{"name":"version_id","in":"path","required":true,"schema":{"type":"integer","title":"Version Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StageUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelVersionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/search":{"get":{"tags":["search"],"summary":"Search Models","operationId":"search_models_api_v1_search_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search query for model name or description","title":"Q"},"description":"Search query for model name or description"},{"name":"team_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Filter by team ID","title":"Team Id"},"description":"Filter by team ID"},{"name":"framework","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/Framework"},{"type":"null"}],"description":"Filter by framework","title":"Framework"},"description":"Filter by framework"},{"name":"stage","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/Stage"},{"type":"null"}],"description":"Filter by stage","title":"Stage"},"description":"Filter by stage"},{"name":"tags","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by tags (comma-separated, format: key:value)","title":"Tags"},"description":"Filter by tags (comma-separated, format: key:value)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ModelList"},"title":"Response Search Models Api V1 Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/import/scan":{"post":{"tags":["import"],"summary":"Scan Legacy Models","operationId":"scan_legacy_models_api_v1_import_scan_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Framework":{"type":"string","enum":["pytorch","tensorflow","sklearn","xgboost","lightgbm","catboost","other"],"title":"Framework"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ModelCreate":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name"],"title":"ModelCreate"},"ModelList":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"team_id":{"type":"integer","title":"Team Id"}},"type":"object","required":["id","name","description","team_id"],"title":"ModelList"},"ModelResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"team_id":{"type":"integer","title":"Team Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","name","description","team_id","created_at","updated_at"],"title":"ModelResponse"},"ModelVersionCreate":{"properties":{"version":{"type":"string","maxLength":100,"minLength":1,"title":"Version"},"artifact_path":{"type":"string","maxLength":1024,"minLength":1,"title":"Artifact Path"},"stage":{"anyOf":[{"$ref":"#/components/schemas/Stage"},{"type":"null"}],"default":"DEVELOPMENT"},"framework":{"anyOf":[{"$ref":"#/components/schemas/Framework"},{"type":"null"}],"default":"other"},"metrics":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Metrics"},"hyperparameters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Hyperparameters"},"dataset_info":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Dataset Info"},"dataset_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Dataset Id"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"created_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created By"},"tag_ids":{"anyOf":[{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Tag Ids"}},"type":"object","required":["version","artifact_path"],"title":"ModelVersionCreate"},"ModelVersionList":{"properties":{"id":{"type":"integer","title":"Id"},"version":{"type":"string","title":"Version"},"model_id":{"type":"integer","title":"Model Id"},"stage":{"$ref":"#/components/schemas/Stage"},"framework":{"$ref":"#/components/schemas/Framework"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","version","model_id","stage","framework","created_at"],"title":"ModelVersionList"},"ModelVersionResponse":{"properties":{"id":{"type":"integer","title":"Id"},"version":{"type":"string","title":"Version"},"model_id":{"type":"integer","title":"Model Id"},"artifact_path":{"type":"string","title":"Artifact Path"},"stage":{"$ref":"#/components/schemas/Stage"},"framework":{"$ref":"#/components/schemas/Framework"},"metrics":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Metrics"},"hyperparameters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Hyperparameters"},"dataset_info":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Dataset Info"},"dataset_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Dataset Id"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"created_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Created By"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","version","model_id","artifact_path","stage","framework","metrics","hyperparameters","dataset_info","dataset_id","description","created_by","created_at","updated_at"],"title":"ModelVersionResponse"},"ScanRequest":{"properties":{"path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Path"}},"type":"object","title":"ScanRequest"},"ScanResponse":{"properties":{"scanned":{"type":"integer","title":"Scanned"},"imported":{"type":"integer","title":"Imported"},"skipped":{"type":"integer","title":"Skipped"},"errors":{"items":{"type":"string"},"type":"array","title":"Errors"}},"type":"object","required":["scanned","imported","skipped","errors"],"title":"ScanResponse"},"Stage":{"type":"string","enum":["DEVELOPMENT","STAGING","PRODUCTION","ARCHIVED"],"title":"Stage"},"StageUpdate":{"properties":{"stage":{"$ref":"#/components/schemas/Stage"}},"type":"object","required":["stage"],"title":"StageUpdate"},"TeamCreate":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name"],"title":"TeamCreate"},"TeamList":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["id","name","description"],"title":"TeamList"},"TeamResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","name","description","created_at","updated_at"],"title":"TeamResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}